+++ CHANGELOG.md
@@ -17,6 +17,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#1486]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/1486
 [#1686]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/1686
 
+## [5.4 patch 1]
+
+### Fixed
+
+- Importing settings from shared URL doesn't work [#3517].
+
+[5.4 patch 1]: https://github.com/AdguardTeam/AdguardBrowserExtension/compare/v5.4.2.0...HEAD
+[#3517]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3517
+
 ## [5.4]
 
 ### Added
@@ -63,7 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Blocked requests in cross-domain iframes were not counted in the extension badge [#3446].
 - Custom filter subscription accepts URLs that return HTML pages instead of filter lists [#3501].
 
-[5.4]: https://github.com/AdguardTeam/AdguardBrowserExtension/compare/v5.3.1.7...HEAD
+[5.4]: https://github.com/AdguardTeam/AdguardBrowserExtension/compare/v5.3.1.7...v5.4.1.4
 [#2485]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2485
 [#2944]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2944
 [#3255]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3255